+++ packages/arch3-core/src/modules/rewards/queries.ts
@@ -10,7 +10,6 @@ import {
   QueryRewardsRecordsResponse
 } from '@archwayhq/arch3-proto/build/archway/rewards/v1/query';
 import { createPagination, QueryClient } from '@cosmjs/stargate';
-import Long from 'long';
 
 export interface RewardsExtension {
   readonly rewards: {
@@ -41,7 +40,7 @@ export function setupRewardsExtension(base: QueryClient): RewardsExtension {
       blockRewardsTracking: () => queryService.blockRewardsTracking(),
       rewardsPool: () => queryService.rewardsPool(),
       estimateTxFees: (gasLimit: number, contractAddress: string) => queryService.estimateTxFees({
-        gasLimit: Long.fromNumber(gasLimit),
+        gasLimit: BigInt(gasLimit),
         contractAddress,
       }),
       rewardsRecords: (rewardsAddress: string, paginationKey?: Uint8Array) => queryService.rewardsRecords({