+++ packages/arch3-core/src/signingarchwayclient.ts
@@ -19,14 +19,13 @@ import {
   StdFee
 } from '@cosmjs/stargate';
 import {
+  CometClient,
   HttpBatchClient,
   HttpBatchClientOptions,
   RpcClient,
-  Tendermint37Client,
-  TendermintClient
+  connectComet,
 } from '@cosmjs/tendermint-rpc';
 import _ from 'lodash';
-import Long from 'long';
 
 import { createRewardsAminoConverters, RewardsMsgEncoder, rewardsTypes } from './modules';
 import { IArchwayQueryClient, createArchwayQueryClient } from './queryclient';
@@ -39,6 +38,7 @@ import {
   RewardsPool,
   RewardsRecord
 } from './types';
+import { connectToRpcClient } from './utils';
 
 export interface SigningArchwayClientOptions extends SigningCosmWasmClientOptions {
   /**
@@ -76,9 +76,9 @@ export interface TxResult {
   /** Transaction events. */
   readonly events: readonly Event[];
   /** Amount of gas sent with the transaction. */
-  readonly gasWanted: number;
+  readonly gasWanted: bigint;
   /** Amount of gas consumed by the transaction. */
-  readonly gasUsed: number;
+  readonly gasUsed: bigint;
 }
 
 /**
@@ -131,7 +131,7 @@ export class SigningArchwayClient extends SigningCosmWasmClient implements IArch
   private readonly archwayQueryClient: IArchwayQueryClient;
   private readonly gasAdjustment: number;
 
-  protected constructor(tmClient: TendermintClient | undefined, signer: OfflineSigner, options: SigningArchwayClientOptions) {
+  protected constructor(cometClient: CometClient | undefined, signer: OfflineSigner, options: SigningArchwayClientOptions) {
     const {
       registry = new Registry([...defaultRegistryTypes, ...wasmTypes, ...rewardsTypes]),
       aminoTypes = new AminoTypes({
@@ -142,9 +142,9 @@ export class SigningArchwayClient extends SigningCosmWasmClient implements IArch
       gasAdjustment = defaultGasAdjustment,
     } = options;
 
-    super(tmClient, signer, { ...options, registry, aminoTypes });
+    super(cometClient, signer, { ...options, registry, aminoTypes });
 
-    this.archwayQueryClient = createArchwayQueryClient(tmClient);
+    this.archwayQueryClient = createArchwayQueryClient(cometClient);
     this.gasAdjustment = gasAdjustment;
   }
 
@@ -155,16 +155,14 @@ export class SigningArchwayClient extends SigningCosmWasmClient implements IArch
    * @param signer - The transaction signer configuration.
    * @param options - Options for the signing client.
    * @returns A {@link SigningArchwayClient} connected to the endpoint.
-   *
-   * @see {@link SigningArchwayClient.createWithSigner} if you need Tendermint 0.37 support.
    */
   public static override async connectWithSigner(
     endpoint: string | HttpEndpoint,
     signer: OfflineSigner,
     options: SigningArchwayClientOptions = {},
   ): Promise<SigningArchwayClient> {
-    const tmClient = await Tendermint37Client.connect(endpoint);
-    return SigningArchwayClient.createWithSigner(tmClient, signer, options);
+    const cometClient = await connectComet(endpoint);
+    return SigningArchwayClient.createWithSigner(cometClient, signer, options);
   }
 
   /**
@@ -178,8 +176,6 @@ export class SigningArchwayClient extends SigningCosmWasmClient implements IArch
    * @returns A {@link SigningArchwayClient} connected to the endpoint.
    *
    * @remarks This factory method doesn't support WebSocket endpoints.
-   *
-   * @see {@link SigningArchwayClient.createWithSigner} if you need Tendermint 0.37 support.
    */
   public static async connectWithSignerAndBatchClient(
     endpoint: string | HttpEndpoint,
@@ -188,25 +184,25 @@ export class SigningArchwayClient extends SigningCosmWasmClient implements IArch
     batchClientOptions?: Partial<HttpBatchClientOptions>
   ): Promise<SigningArchwayClient> {
     const rpcClient: RpcClient = new HttpBatchClient(endpoint, batchClientOptions);
-    const tmClient = await Tendermint37Client.create(rpcClient);
-    return SigningArchwayClient.createWithSigner(tmClient, signer, options);
+    const cometBatchClient = await connectToRpcClient(rpcClient);
+    return SigningArchwayClient.createWithSigner(cometBatchClient, signer, options);
   }
 
   /**
-   * Creates an instance from a manually created Tendermint client.
+   * Creates an instance from a manually created CometBFT client.
    *
-   * @param tmClient - A Tendermint client for a given endpoint.
+   * @param cometClient - A CometBFT client for a given endpoint.
    * @param signer - The transaction signer configuration.
    * @param options - Options for the signing client.
    * @returns A {@link SigningArchwayClient} connected to the endpoint.
    */
   /* eslint-disable-next-line @typescript-eslint/require-await */
   public static override async createWithSigner(
-    tmClient: TendermintClient,
+    cometClient: CometClient,
     signer: OfflineSigner,
     options: SigningArchwayClientOptions = {},
   ): Promise<SigningArchwayClient> {
-    return new SigningArchwayClient(tmClient, signer, options);
+    return new SigningArchwayClient(cometClient, signer, options);
   }
 
   /**
@@ -351,7 +347,7 @@ export class SigningArchwayClient extends SigningCosmWasmClient implements IArch
     const message = RewardsMsgEncoder.withdrawRewards({
       rewardsAddress,
       recordsLimit: {
-        limit: Long.fromNumber(limit),
+        limit: BigInt(limit),
       }
     });
     const response = await this.assertSignAndBroadcast(senderAddress, [message], fee, memo);